# specs/001-pipeline-intelligence/contracts/api.yaml
openapi: 3.1.0
info:
  title: HubSpot Pipeline Intelligence API
  version: 1.0.0
  description: >
    Backend REST API for the Pipeline Intelligence system.
    Handles recommendations, deals, admin configuration,
    Slack interactions, audit trails, and system health.
servers:
  - url: /api/v1
    description: API v1 prefix
security:
  - oidcAuth: []
components:
  securitySchemes:
    oidcAuth:
      type: openIdConnect
      openIdConnectUrl: https://accounts.google.com/.well-known/openid-configuration
  schemas:
...
    UserConfig:
      type: object
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
        display_name:
          type: string
        role:
          type: string
          enum: [admin, revops, sales_user]
        hubspot_owner_id:
          type: string
          nullable: true
        slack_user_id:
          type: string
          nullable: true
        is_active:
          type: boolean
        gmail_watch_active:
          type: boolean
        created_at:
          type: string
          format: date-time
    CreateUserRequest:
      type: object
      required:
        - email
        - display_name
        - role
      properties:
        email:
          type: string
          format: email
        display_name:
          type: string
        role:
          type: string
          enum: [admin, revops, sales_user]
        hubspot_owner_id:
          type: string
          nullable: true
        slack_user_id:
          type: string
          nullable: true
    PipelineConfig:
      type: object
      properties:
        id:
          type: string
          format: uuid
        hubspot_pipeline_id:
          type: string
        pipeline_name:
          type: string
        stages:
          type: array
          items:
            type: object
            properties:
              id:
                type: string
              name:
                type: string
              order:
                type: integer
        intent_to_stage_rules:
          type: object
          additionalProperties:
            type: string
        is_active:
          type: boolean
        recommendation_threshold:
          type: number
        logging_threshold:
          type: number
        approval_timeout_hours:
          type: integer
        max_snoozes:
          type: integer
    UpdatePipelineConfigRequest:
      type: object
      properties:
        intent_to_stage_rules:
          type: object
          additionalProperties:
            type: string
        recommendation_threshold:
          type: number
          minimum: 0
          maximum: 1
        logging_threshold:
          type: number
          minimum: 0
          maximum: 1
        approval_timeout_hours:
          type: integer
          minimum: 1
        max_snoozes:
          type: integer
          minimum: 0
        is_active:
          type: boolean
    PromptVersion:
      type: object
      properties:
        id:
          type: string
          format: uuid
        version:
          type: string
        prompt_template:
          type: string
        system_prompt:
          type: string
        intent_categories:
          type: array
          items:
            type: string
        is_active:
          type: boolean
        created_by:
          type: string
          format: uuid
          nullable: true
        notes:
          type: string
          nullable: true
        created_at:
          type: string
          format: date-time
    CreatePromptVersionRequest:
      type: object
      required:
        - version
        - prompt_template
        - system_prompt
        - intent_categories
      properties:
        version:
          type: string
        prompt_template:
          type: string
        system_prompt:
          type: string
        intent_categories:
          type: array
          items:
            type: string
        notes:
          type: string
          nullable: true
...
paths:
...
  /admin/users:
    get:
      operationId: listUsers
      summary: List all system users (admin only)
      tags: [Admin]
      responses:
        '200':
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/UserConfig'
    post:
      operationId: createUser
      summary: Add a new user
      tags: [Admin]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateUserRequest'
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserConfig'
  /admin/users/{id}:
    patch:
      operationId: updateUser
      summary: Update user configuration
      tags: [Admin]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                display_name:
                  type: string
                role:
                  type: string
                  enum: [admin, revops, sales_user]
                hubspot_owner_id:
                  type: string
                slack_user_id:
                  type: string
                is_active:
                  type: boolean
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserConfig'

    delete:
      operationId: deactivateUser
      summary: Deactivate a user (soft delete)
      tags: [Admin]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: User deactivated

  # ── Admin: Pipelines ─────────────────────────────────────────
...
  /admin/pipelines/{id}:
    get:
      operationId: getPipelineConfig
      summary: Get pipeline configuration
      tags: [Admin]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PipelineConfig'

    patch:
      operationId: updatePipelineConfig
      summary: Update pipeline configuration
      tags: [Admin]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdatePipelineConfigRequest'
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PipelineConfig'

  # ── Admin: Prompts ───────────────────────────────────────────
  /admin/prompts:
    get:
      operationId: listPromptVersions
      summary: List all prompt versions
      tags: [Admin]
      responses:
        '200':
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/PromptVersion'

    post:
      operationId: createPromptVersion
      summary: Create a new prompt version
      tags: [Admin]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatePromptVersionRequest'
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PromptVersion'
  /admin/prompts/{id}/activate:
    post:
      operationId: activatePromptVersion
      summary: Activate a prompt version (deactivates current active)
      tags: [Admin]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PromptVersion'
  /admin/prompts/{id}/rollback:
    post:
      operationId: rollbackPromptVersion
      summary: Rollback to a previous prompt version
      tags: [Admin]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PromptVersion'

  # ── Auth ─────────────────────────────────────────────────────
...